style: clippy

+++ src/auth/token/user_auth.rs
@@ -57,7 +57,7 @@ mod tests {
     #[test]
     fn returns_true_for_expired_token() -> anyhow::Result<()> {
         let id = Uuid::new_v4();
-        let exp = (Utc::now().timestamp() - 60000_i64) as i64;
+        let exp = Utc::now().timestamp() - 60000_i64;
         let claim = TokenClaim::new(id, exp, TokenType::UserAuth, TokenRole::User, None);
         let token = UserAuthToken::try_new(claim)?;
